Move Entry/Weak into new module
